pyproject.toml
[tool.poetry]
name = "dbt-autodoc"
version = "0.1.0"
description = "`dbt-autodoc` is a Singer target for mssql, built with the Meltano SDK for Singer Targets."
authors = ["Henning Holgersen"]
keywords = [
    "documentation",
    "dbt",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.14,>=3.8.0"
openai = ">1.1.0"
click = "^8.1.3"
"ruamel.yaml" = "^0.18.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^22.9"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "target_mssql"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
dbt-autodoc = 'dbt_autodoc.cli:autodoc'
